+++ trakka/components/project/funcs.py
@@ -8,16 +8,16 @@ from trakka.utils.helpers.project import get_project_by_abbrev
 from trakka.utils.misc import logger_wraps
 from trakka.utils.output import get_viewtype_columns
 from trakka.utils.paths import PROJECT_PATH
-from trakka.utils.paths import SET_TYPE
 from trakka.utils.paths import SET_DASHBOARD
 from trakka.utils.paths import ASSIGNED_DASHBOARD
 
 compact_fields = [
     "projectId",        # Project ID
     "abbreviation",     # Abbreviation or short name
-    "type",             # Type for the project
+    "label",            # Type for the project
     "isActive",         # Active status
-    "name"              # Full name of the project
+    "name",             # Full name of the project
+    "status",  # The current status of the project
 ]
 
 more_fields = [
@@ -25,7 +25,7 @@ more_fields = [
     "projectId",        # Project ID
     "abbreviation",     # Abbreviation or short name
     "clientType",       # the 'secret' client type
-    "type",             # Type for the project
+    "label",             # Type for the project
     "isActive",         # Active status
     "name",             # Project name
     "description",      # Description of the project
@@ -43,7 +43,7 @@ def add_project(
         description: str,
         org: str,
         dashboard_name: str,
-        project_type: str,
+        project_label: str,
         client_type: str,
         merge_algo: str):
     return api_post(
@@ -53,11 +53,9 @@ def add_project(
             "name": name,
             "description": description,
             "isActive": True,
-            "requestingOrg": {
-                "abbreviation": org
-            },
+            "requestingOrg": org,
             "dashboardName": dashboard_name,
-            "type": project_type,
+            "label": project_label,
             "clientType": client_type,
             "mergeAlgorithm": merge_algo, 
        }
@@ -69,12 +67,12 @@ def update_project(
         project_abbreviation: str,
         name: str,
         description: str,
-        is_active: bool,
         org: str,
         dashboard_name: str,
-        project_type: str,
+        project_label: str,
         client_type: str,
-        merge_algorithm: str
+        merge_algorithm: str,
+        status: str,
 ):
     project = get_project_by_abbrev(project_abbreviation)
     
@@ -83,36 +81,34 @@ def update_project(
         **{k: project[k] for k in [
             'name',
             'description',
-            'isActive',
             'requestingOrg',
             'dashboardName',
-            'type',
+            'label',
             'clientType',
-            'mergeAlgorithm'
+            'mergeAlgorithm',
+            'status',
         ]},
     }
     
     if project['requestingOrg'] is None:
-        put_project['requestingOrg'] = {'abbreviation': None}
+        put_project['requestingOrg'] = None
 
     if name is not None:
         put_project['name'] = name
     if description is not None:
         put_project['description'] = description
-    if is_active is not None:
-        put_project['isActive'] = is_active
     if org is not None:
-        put_project['requestingOrg'] = {
-            "abbreviation": org
-        }
+        put_project['requestingOrg'] = org
     if dashboard_name is not None:
         put_project['dashboardName'] = dashboard_name
-    if project_type is not None:
-        put_project['type'] = project_type
+    if project_label is not None:
+        put_project['label'] = project_label
     if client_type is not None:
         put_project['clientType'] = client_type
     if merge_algorithm is not None:
         put_project['mergeAlgorithm'] = merge_algorithm
+    if status is not None:
+        put_project['status'] = status
         
     return api_put(
         path=f"{PROJECT_PATH}/{project_abbreviation}",
@@ -141,12 +137,6 @@ def list_projects(view_type: str, out_format: str):
 def get_dashboard(project_abbreviation: str, out_format: str):
     joined_path = '/'.join([PROJECT_PATH, ASSIGNED_DASHBOARD, project_abbreviation])
     call_get_and_print(joined_path, out_format)
-
-    
-@logger_wraps() 
-def set_project_type(abbrev: str, project_type: str):
-    path = '/'.join([PROJECT_PATH, abbrev, SET_TYPE])
-    api_patch(path, data=project_type,)
 
 @logger_wraps()
 def enable_project(abbrev: str):