+++ trakka/main.py
@@ -6,13 +6,16 @@ import click
 from click.core import Context
 from loguru import logger
 
+from trakka.utils.config import get_server_info_or_create
+from trakka.utils.privilege import TENANT_RESOURCE
+
 from trakka.utils.context import CxtKey
 from trakka.utils.context import TrakkaCxt
 from trakka.components.admin import admin
 from trakka.components.auth import auth
 from trakka.components.user import user
 from trakka.components.org import org
-from trakka.components.log import log
+from trakka.components.log import log_subcommands
 from trakka.components.project import project
 from trakka.components.tree import tree
 from trakka.components.metadata import metadata
@@ -38,7 +41,8 @@ from trakka.utils.logger import setup_logger
 from trakka.utils.logger import LOG_LEVEL_INFO
 from trakka.utils.logger import LOG_LEVELS
 from trakka.utils.cmd_filter import USER, show_admin_cmds
-from trakka.utils.version import check_version
+from trakka.utils.cmd_filter import show_admin_cmds
+from trakka.utils.version import check_version, warn_if_austrakka
 
 
 CONTEXT_SETTINGS = {"help_option_names": HELP_OPTS}
@@ -154,8 +158,17 @@ def cli(
         CxtKey.CMD_SET.value: cmd_set,
     }
     setup_logger(log_level, log_var)
+    warn_if_austrakka()
     if not skip_version_check:
-        check_version(VERSION)
+        server_info = get_server_info_or_create(
+            uri,
+            skip_cert_verify,
+        )
+        if server_info is None:
+            check_version(VERSION)
+        else:
+            logger.debug("Not checking for new CLI updates as server is not"
+                         + " rolling.")
     TrakkaCxt.check_deprecated_env_vars()
 
 
@@ -176,7 +189,7 @@ def get_cli():
     cli.add_command(field)
     cli.add_command(fieldtype)
     cli.add_command(iam) if show_admin_cmds() else None
-    cli.add_command(log) if show_admin_cmds() else None
+    cli.add_command(log_subcommands(TENANT_RESOURCE)) if show_admin_cmds() else None
     return cli