+++ CHANGELOG.md
@@ -5,6 +5,77 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog],
 and this project adheres to [Semantic Versioning].
 
+## [0.89.0] - 2026-07-14
+
+### Added
+- `proforma rm-attach` command to remove attached proforma templates.
+- `--status` parameter to update project status in `project update`.
+
+### Changed
+- `group field [add|remove|list]` -> `org field [add|remove|list]`. Now operates on an organisation rather than a group.
+- `metadata list` -> `org metadata get`. Now operates on an organisation rather than a group.
+- Clean up contig name generated fasta file for `seq sync` command.
+- `--type` -> `--label` for `project [add|update]`.
+
+### Removed
+- `project set-type`. This can be set using `project [add|update]`.
+- `[--is-active|--not-active]` from `project update`: use `project [enable|disable]` instead. 
+
+### Fixed
+- Fixed displaying of columns for `field list` using `--view-type more`.
+
+## [0.88.1] - 2026-06-19
+
+### Added
+- The `proforma share` command can now take a `--project` parameter to share a proforma with a project, as an
+alternative to using the `--group-name` parameter to share with the project group. This syntax is more consistent with
+other commands and will be used in the future. The `--group-name` parameter is still supported in this release.
+
+## [0.88.0] - 2026-06-17
+
+## Changed
+- `user move-org` is now `user transfer`.
+- `sample chown` is now `sample transfer`, with parameters consistent with `user transfer`.
+
+## [0.87.1] - 2026-06-16
+
+### Changed
+- Reverted renaming of provision commands.
+
+## [0.87.0] - 2026-06-15
+
+### Changed
+- Obsolete project provision commands removed.
+- Will not warn about out-of-date CLI version if the user is targeting a non-rolling Trakka instance.
+
+## [0.86.0] - 2026-06-01
+
+### Changed
+- Update to match change to the API. This is a backwards-incompatible change which requires a CLI version update. 
+This specifically affects the `project tree add` command.
+
+## [0.85.0] - 2026-05-20
+
+### Added 
+- `project document [add|disable|download|enable|list|list-all|update]` commands.
+- `user move-org` command.
+
+## [0.84.1] - 2026-05-04
+
+### Changed
+- `sample chown` is no longer an admin-only command.
+
+## [0.84.0] - 2026-04-21
+
+### Removed
+- `-rt/--record-type` and `-id` options from `iam privilege`; for non-system privileges use `[org|project] privilege`. 
+- `-rt/--record-type` and `-id` options from `log`; for non-system logs use `[org|project] log`. 
+
+## [0.83.1] - 2026-04-20
+
+### Changed
+- Improved performance of `admin rlog regen-bulk` command.
+
 ## [0.83.0] - 2026-04-14
 
 ### Added